🏗️ Move everything into a single file
