--- lamin_cli/compute/modal/__init__.py
@@ -1,2 +1,0 @@
-from ._modal_runner import Runner
-from ._modal_script_function import run_script_from_path
